+++ NEWS.md
@@ -4,6 +4,8 @@
 
 - added an argument `quiet` to `plot_crop()` (thanks, @WastlM, #1034)
 
+- support for !expr in knit_params (correct 'type' field for Date and POSIXct objects generated via !eval; add 'expr' field to flag when a parameter's value was the result of an R expression)
+
 ## MAJOR CHANGES
 
 - the two hook functions `hook_rgl()` and `hook_webgl()` have been moved from **knitr** to the **rgl** package (v0.95.1247) (thanks, @dmurdoch)